Restructure and move Pupil and Student to own file
