+++ src/algo/addon/code/src/types.h
@@ -66,7 +66,7 @@ namespace CS {
     };
 
     ///Interface for nodes. It serves as a base class for students and pupils.
-    class NodeIF {
+    class Node {
     public:
         const ID id;
         
@@ -77,118 +77,10 @@ namespace CS {
         std::vector<UUID> dissolved_matches_with;
         double waiting_days = 0; //Want have also fractions of a day, therefore double..
 
-        inline explicit NodeIF(ID id): id(id){ }
+        inline explicit Node(ID id): id(id){ }
     };
 
-    class Pupil;
-
-    class CollegeStudent : public NodeIF {
-    public:
-        using NodeIF::id;
-        std::vector<OfferedSubject> offered_subjects;
-        unsigned number_of_possible_matches{0u};
-        
-        inline explicit CollegeStudent(ID id);
-
-        [[nodiscard]] inline bool accepts(const Pupil & pupil) const {
-            if (_acceptance_function) {
-                return _acceptance_function(pupil);
-            }
-            return false;
-        }
-    private:
-        std::function<bool(const Pupil &)> _acceptance_function;
-    };
-
-    class Pupil : public NodeIF {
-    public:
-        using NodeIF::id;
-
-        std::vector<RequestedSubject> requested_subjects;
-        double matching_priority{0.};
-        Grade grade = MIN_POSSIBLE_GRADE;
-        
-        inline explicit Pupil(ID id) : NodeIF(id) {}
-
-    private:
-        std::function<bool(const CollegeStudent &)> _acceptance_function;
-    };
-
-    class NodeContainer {
-    public:
-        [[nodiscard]] inline const CollegeStudent & college_student(ID id) const {
-            if (id >= _students.size()) {
-                throw std::out_of_range("Tried to access a student with an invalid ID");
-            }
-            return _students[id];
-        }
-
-        [[nodiscard]] inline CollegeStudent & college_student(ID id) {
-            if (id >= _students.size()) {
-                throw std::out_of_range("Tried to access a student with an invalid ID");
-            }
-            return _students[id];
-        }
-
-        [[nodiscard]] inline const Pupil & pupil(ID id) const {
-            if (id >= _pupils.size()) {
-                throw std::out_of_range("Tried to access a pupil with an invalid ID");
-            }
-            return _pupils[id];
-        }
-
-        [[nodiscard]] inline Pupil & pupil(ID id) {
-            if (id >= _pupils.size()) {
-                throw std::out_of_range("Tried to access a pupil with an invalid ID");
-            }
-            return _pupils[id];
-        }
-
-        [[nodiscard]] inline  const std::vector<Pupil> & pupils() const {
-            return _pupils;
-        }
-
-        [[nodiscard]] inline std::vector<Pupil> & pupils(){
-            return _pupils;
-        }
-
-        [[nodiscard]] inline  const std::vector<CollegeStudent> & college_students() const {
-            return _students;
-        }
-
-        [[nodiscard]] inline std::vector<CollegeStudent> & college_students(){
-            return _students;
-        }
-
-        inline void create_pupils(unsigned num) {
-            _pupils.reserve(_pupils.size() + num);
-            auto const old_size = _pupils.size();
-            for (auto id = old_size; id < old_size + num; ++id) {
-                _pupils.emplace_back(id);
-            }
-        }
-
-        inline void create_college_students(unsigned num) {
-            _students.reserve(_students.size() + num);
-            auto const old_size = _students.size();
-            for (auto id = old_size; id < old_size + num; ++id) {
-                _students.emplace_back(id);
-            }
-        }
-
-        [[nodiscard]] inline std::uint32_t size() const {
-            return _students.size() + _pupils.size();
-        }
-
-
-    private:
-        std::vector<CollegeStudent> _students;
-        std::vector<Pupil> _pupils;
-    };
-
-    CollegeStudent::CollegeStudent(ID id) : NodeIF(id) {}
-
-
+    
 
     inline double get_day_difference_from_today(const std::string &date) {
         static constexpr unsigned SECONDS_IN_A_DAY = 86400;
